NavGraph for AuthorizationActivity. Registration/authorization features separated from MainActivity
