+++ app/src/main/java/com/example/students/App.kt
@@ -6,9 +6,9 @@ import com.example.students.di.network.commonModule
 import com.example.students.di.network.networkModule
 import com.example.students.features.feed.di.feedPageModule
 import com.example.students.features.form.di.formModule
-import com.example.students.features.login.di.loginModule
+import com.example.students.features.auth.login.di.loginModule
 import com.example.students.features.otp.di.otpModule
-import com.example.students.features.registration.di.registrationModule
+import com.example.students.features.auth.registration.di.registrationModule
 import com.yandex.mapkit.MapKitFactory
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
