+++ app/src/main/java/com/example/students/features/auth/AuthorizationActivity.kt
@@ -1,0 +1,16 @@
+package com.example.students.features.auth
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.students.databinding.ActivityAuthorizationBinding
+
+class AuthorizationActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityAuthorizationBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityAuthorizationBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+}
