+++ app/src/main/java/com/example/students/features/auth/login/data/model/LoginRequest.kt
@@ -1,0 +1,12 @@
+package com.example.students.features.auth.login.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class LoginRequest(
+    @SerializedName("phone")
+    val phone: String,
+    @SerializedName("password")
+    val password: String,
+    @SerializedName("code")
+    val code: String,
+)
