+++ app/src/main/java/com/example/students/features/auth/login/data/model/LoginResponse.kt
@@ -1,0 +1,12 @@
+package com.example.students.features.auth.login.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class LoginResponse(
+    @SerializedName("access_token")
+    val accessToken: String,
+    @SerializedName("expires_in")
+    val expiresIn: Int,
+    @SerializedName("token_type")
+    val tokenType: String
+)
