+++ app/src/main/java/com/example/students/features/auth/registration/data/model/RegistrationRequest.kt
@@ -1,0 +1,10 @@
+package com.example.students.features.auth.registration.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class RegistrationRequest(
+    @SerializedName("phone")
+    val phone: String,
+    @SerializedName("password")
+    val password: String
+)
