+++ app/src/main/java/com/example/students/features/auth/registration/presentation/RegistrationScreenState.kt
@@ -1,0 +1,8 @@
+package com.example.students.features.auth.registration.presentation
+
+open class RegistrationScreenState {
+    object Loading: RegistrationScreenState()
+    class ErrorLoaded(val message: String) : RegistrationScreenState()
+    class RegistrationSuccess: RegistrationScreenState()
+    class RegistrationFailed: RegistrationScreenState()
+}
