+++ app/src/main/java/com/example/students/features/form/domain/repository/FormRepository.kt
@@ -2,7 +2,6 @@ package com.example.students.features.form.domain.repository
 
 import com.example.students.features.form.data.model.ProfileRequest
 import com.example.students.features.form.data.model.ProfileResponse
-import com.example.students.features.login.data.model.LoginRequest
 import com.example.students.utils.Resource
 
 interface FormRepository {
