--- app/src/main/java/com/example/students/features/login/data/model/LoginRequest.kt
@@ -1,12 +1,0 @@
-package com.example.students.features.login.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class LoginRequest(
-    @SerializedName("phone")
-    val phone: String,
-    @SerializedName("password")
-    val password: String,
-    @SerializedName("code")
-    val code: String,
-)
