--- app/src/main/java/com/example/students/features/login/presentation/LoginScreenState.kt
@@ -1,8 +1,0 @@
-package com.example.students.features.login.presentation
-
-open class LoginScreenState {
-    object Loading: LoginScreenState()
-    class ErrorLoaded(val message: String) : LoginScreenState()
-    object LoginSuccess: LoginScreenState()
-    object LoginFailed: LoginScreenState()
-}
