--- app/src/main/res/navigation/nav_graph.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/splashFragment">
-
-    <fragment
-        android:id="@+id/navigation_login"
-        android:name="com.example.students.features.login.presentation.LoginFragment"
-        android:label="LoginFragment"
-        tools:layout="@layout/fragment_login">
-        <action
-            android:id="@+id/action_navigation_login_to_registrationFragment"
-            app:destination="@id/navigation_registration" />
-        <action
-            android:id="@+id/action_navigation_login_to_mainFragment"
-            app:destination="@id/feedFragment"
-            app:popUpTo="@+id/main_nav_graph"
-            app:popUpToInclusive="true" />
-    </fragment>
-    <fragment
-        android:id="@+id/navigation_registration"
-        android:name="com.example.students.features.registration.presentation.RegistrationFragment"
-        android:label="RegistrationFragment"
-        tools:layout="@layout/fragment_registration">
-        <action
-            android:id="@+id/action_registrationFragment_to_navigation_login"
-            app:destination="@id/navigation_login" />
-        <action
-            android:id="@+id/action_registrationFragment_to_formFragment"
-            app:destination="@id/navigation_form" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_form"
-        android:name="com.example.students.features.form.presentation.FormFragment"
-        android:label="FormFragment"
-        tools:layout="@layout/fragment_form">
-        <action
-            android:id="@+id/action_navigation_form_to_mainFragment"
-            app:destination="@id/feedFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/feedFragment"
-        android:name="com.example.students.features.feed.presentation.FeedFragment"
-        android:label="fragment_feed"
-        tools:layout="@layout/fragment_feed" />
-    <fragment
-        android:id="@+id/mapFragment"
-        android:name="com.example.students.features.map.presentation.MapFragment"
-        android:label="fragment_map"
-        tools:layout="@layout/fragment_map" />
-    <fragment
-        android:id="@+id/splashFragment"
-        android:name="com.example.students.features.splash.SplashFragment"
-        android:label="fragment_splash"
-        tools:layout="@layout/fragment_splash">
-        <action
-            android:id="@+id/action_splashFragment_to_onboardingFragment"
-            app:destination="@id/onboardingFragment" />
-        <action
-            android:id="@+id/action_splashFragment_to_navigation_login"
-            app:destination="@id/navigation_login" />
-    </fragment>
-    <fragment
-        android:id="@+id/onboardingFragment"
-        android:name="com.example.students.features.onboarding.presentation.OnboardingFragment"
-        android:label="fragment_onboarding"
-        tools:layout="@layout/fragment_onboarding">
-        <action
-            android:id="@+id/action_onboardingFragment_to_navigation_login"
-            app:destination="@id/navigation_login"
-            app:popUpTo="@+id/main_nav_graph"
-            app:popUpToInclusive="true" />
-    </fragment>
-
-</navigation>
